refactor(admin): split dashboard into feature pages
Extract admin routes into dedicated page components with AdminLayout and shared theme form helpers. Remove unused StoreSettingsForm, ProductForm, and kv-helper. Use securityHeadersMiddleware in worker.
