+++ src/pages/admin/ProductsPage.jsx
@@ -1,0 +1,5 @@
+import { ProductWorkspace } from '../../components/admin/ProductWorkspace'
+
+export function ProductsPage() {
+  return <ProductWorkspace />
+}
