+++ src/pages/admin/StoreSettingsPage.jsx
@@ -1,0 +1,5 @@
+import { StoreSettingsEditor } from '../../components/admin/StoreSettingsEditor'
+
+export function StoreSettingsPage() {
+  return <StoreSettingsEditor />
+}
