+++ src/worker.js
@@ -1,6 +1,7 @@
 // Main Cloudflare Worker with Hono framework
 import { Hono } from 'hono'
 import { createCorsMiddleware } from './middleware/cors.js'
+import { securityHeadersMiddleware } from './middleware/securityHeaders.js'
 import { errorHandler } from './middleware/errorHandler.js'
 import { verifyAdminAuth } from './middleware/auth.js'
 import { productLimitMiddleware } from './middleware/productLimit.js'
@@ -8,40 +9,7 @@ import { registerRoutes } from './routes/index.js'
 
 const app = new Hono()
 
-// Security headers middleware (applied to all routes)
-app.use('*', async (c, next) => {
-  await next()
-  
-  // Security headers
-  c.header('X-Content-Type-Options', 'nosniff')
-  c.header('X-Frame-Options', 'DENY')
-  c.header('X-XSS-Protection', '1; mode=block')
-  c.header('Referrer-Policy', 'strict-origin-when-cross-origin')
-  c.header('Permissions-Policy', 'geolocation=(), microphone=(), camera=()')
-  
-  // Content Security Policy
-  const csp = [
-    "default-src 'self'",
-    "script-src 'self' 'unsafe-inline' 'unsafe-eval' https://js.stripe.com",
-    "style-src 'self' 'unsafe-inline'",
-    "img-src 'self' data: https: blob:",
-    "font-src 'self' data:",
-    "connect-src 'self' https://api.stripe.com https://*.stripe.com https://oauth2.googleapis.com https://www.googleapis.com",
-    "frame-src https://js.stripe.com https://hooks.stripe.com",
-    "frame-ancestors 'none'",
-    "base-uri 'self'",
-    "form-action 'self'",
-    "upgrade-insecure-requests"
-  ].join('; ')
-  
-  c.header('Content-Security-Policy', csp)
-  
-  // Strict Transport Security (only on HTTPS)
-  const url = new URL(c.req.url)
-  if (url.protocol === 'https:') {
-    c.header('Strict-Transport-Security', 'max-age=31536000; includeSubDomains; preload')
-  }
-})
+app.use('*', securityHeadersMiddleware)
 
 // CORS middleware (needs env for proper configuration)
 // Note: We create it per-request since env is request-specific
